app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alchemical==0.5.1
    # via -r requirements.in
alembic==1.7.7
    # via flask-migrate
apifairy==1.3.0
    # via -r requirements.in
apispec==5.1.1
    # via apifairy
blinker==1.4
    # via flask-mail
click==8.1.0
    # via flask
faker==13.3.4
    # via -r requirements.in
flask==2.1.0
    # via
    #   -r requirements.in
    #   apifairy
    #   flask-cors
    #   flask-httpauth
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-sqlalchemy
flask-cors==3.0.10
    # via -r requirements.in
flask-httpauth==4.5.0
    # via
    #   -r requirements.in
    #   apifairy
flask-mail==0.9.1
    # via -r requirements.in
flask-marshmallow==0.14.0
    # via
    #   -r requirements.in
    #   apifairy
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via flask-migrate
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.1
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.15.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-sqlalchemy==0.28.0
    # via -r requirements.in
packaging==21.3
    # via
    #   marshmallow
    #   webargs
pyjwt==2.3.0
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
python-dateutil==2.8.2
    # via faker
python-dotenv==0.20.0
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-cors
    #   flask-marshmallow
    #   python-dateutil
sqlalchemy==1.4.32
    # via
    #   alchemical
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
webargs==8.1.0
    # via apifairy
werkzeug==2.1.0
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
